refactor: move asset pipelines to asset modules
