+++ Cargo.toml
@@ -22,16 +22,13 @@ axum = { version = "0.6", features = ["ws"] }
 bytes = "1"
 clap = { version = "4", features = ["derive", "env"] }
 console = "0.15"
-dunce = "1"
 envy = "0.4"
-fs_extra = "1"
 futures-util = { version = "0.3", default-features = false, features = [
     "sink",
 ] }
 local-ip-address = "0.5.1"
 nipper = "0.1"
 notify = "4"
-once_cell = "1"
 open = "5"
 reqwest = { version = "0.11", default-features = false, features = [
     "rustls-tls",
@@ -54,9 +51,6 @@ tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 trunk-util = { version = "0.1", path = "crates/util" }
 trunk-tools = { version = "0.1", path = "crates/tools" }
 trunk-pipelines = { version = "0.1", path = "crates/pipelines" }
-
-[dev-dependencies]
-tempfile = "3"
 
 [workspace]
 members = ["crates/*"]
