+++ crates/pipelines/src/assets/copy_file/tests.rs
@@ -4,9 +4,9 @@ use std::sync::Arc;
 
 use anyhow::{Context, Result};
 
+use crate::assets::{CopyFile, CopyFileConfig, Pipeline};
 // use crate::config::RtcBuild;
 use crate::util::ATTR_HREF;
-use crate::{CopyFile, CopyFileConfig, Pipeline};
 
 struct CopyFileTestConfig {
     output_dir: PathBuf,
