+++ crates/pipelines/src/assets/css.rs
@@ -9,9 +9,9 @@ use futures_util::FutureExt;
 use nipper::Document;
 use tokio::task::JoinHandle;
 
-use super::asset_file::AssetFile;
+use super::{Output, Pipeline};
+use crate::asset_file::AssetFile;
 use crate::util::{trunk_id_selector, Attrs, ErrorReason, Result, ResultExt, ATTR_HREF};
-use crate::{Output, Pipeline};
 
 /// A trait that indicates a type can be used as config type for css pipeline.
 pub trait CssConfig {
