+++ crates/pipelines/src/assets/icon.rs
@@ -8,9 +8,9 @@ use futures_util::FutureExt;
 use nipper::Document;
 use tokio::task::JoinHandle;
 
+use super::{Output, Pipeline};
 use crate::asset_file::AssetFile;
 use crate::util::{trunk_id_selector, Attrs, ErrorReason, Result, ResultExt, ATTR_HREF};
-use crate::{Output, Pipeline};
 
 /// A trait that indicates a type can be used as config type for icon pipeline.
 pub trait IconConfig {
