+++ crates/pipelines/src/assets/output.rs
