+++ crates/pipelines/src/assets/rust/cargo.rs
