+++ crates/pipelines/src/assets/rust/config.rs
