+++ crates/pipelines/src/assets/rust/wasm_opt.rs
