+++ src/pipelines/mod.rs
@@ -12,7 +12,7 @@ use nipper::Document;
 use serde::Deserialize;
 use tokio::sync::mpsc;
 use tokio::task::JoinHandle;
-use trunk_pipelines::{
+use trunk_pipelines::assets::{
     CopyDir, CopyDirConfig, CopyDirOutput, CopyFile, CopyFileConfig, CopyFileOutput, Css,
     CssConfig, CssOutput, Icon, IconConfig, IconOutput, Inline, InlineOutput, Js, JsConfig,
     JsOutput, Output, Pipeline, RustApp, RustAppConfig, RustAppOutput, Sass, SassConfig,
